test(richtext): reorganize RichString test files by edit targets (span vs paragraph)
To improve intuitiveness, tests inside `RichStringTest` were extracted based on their conceptual domains. Methods regarding `setAttribute` / `removeAttribute` behavior were moved to `RichStringEditSpanTest`. Existing tests covering `setParagraphAttribute` mapping and snapping were moved to `RichStringEditParagraphTest`.
